requirements.txt
Flask==2.3.3
face-recognition==1.3.0
nudenet>=3.4.2
requests==2.31.0
numpy==1.26.4
Pillow==10.3.0
insightface==0.7.3
onnxruntime
python-dotenv
deepface
tf-keras
torch
torchvision
openai-clip
transformers
scikit-learn
supabase